pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inqtrix"
version = "0.1.0"
description = "Iterative AI research agent with parallel web search, claim verification, and source tiering"
readme = "README.md"
authors = [{ name = "Babak Zandi" }]

license = "Apache-2.0"
requires-python = ">=3.12"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "openai>=1.80,<2.0",
    "langgraph>=0.4,<1.0",
    "cachetools>=5.5,<6.0",
    "pydantic-settings>=2.7,<3.0",
    "python-dotenv>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0",
    "boto3>=1.35",
    "azure-identity>=1.18",
    "azure-ai-projects>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "requests>=2.31",
    "vcrpy>=8.1",
    "pytest-recording>=0.13",
]
ui = [
    "streamlit>=1.55,<2.0",
    "httpx>=0.28,<1.0",
]

[project.scripts]
inqtrix-parity = "inqtrix.parity:main"

[tool.hatch.build.targets.wheel]
packages = ["src/inqtrix"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "vcrpy>=8.1",
    "pytest-recording>=0.13",
]

[tool.pytest.ini_options]
markers = [
    "replay: provider tests using VCR cassettes or botocore.Stubber (offline-replayable)",
]
